test6-FINAL/pyproject.toml
[project]
name = "test2"
version = "0.1.0"
description = "test2 using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "arxiv>=2.2.0",
    "chromadb>=1.1.0",
    "crewai[tools]>=0.193.0,<1.0.0",
    "fastmcp>=2.12.4",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "langchain-nvidia-ai-endpoints>=0.3.18",
    "langchain-ollama>=0.3.10",
    "langchain-text-splitters>=0.3.11",
    "pypdfium2>=4.30.0",
    "rank-bm25>=0.2.2",
]

[project.scripts]
kickoff = "test2.research_flow:kickoff"
run_crew = "test2.research_flow:kickoff"
plot = "test2.research_flow:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
